chore: Move ITransactionDetailsRequest.ts under chrysalis
