+++ client/src/models/api/chrysalis/ITransactionsDetailsRequest.ts
