+++ client/src/services/chrysalis/chrysalisTangleCacheService.ts
@@ -6,9 +6,9 @@ import { asTransactionObject } from "@iota/transaction-converter";
 import { ServiceFactory } from "../../factories/serviceFactory";
 import { TrytesHelper } from "../../helpers/trytesHelper";
 import { ISearchResponse } from "../../models/api/chrysalis/ISearchResponse";
+import { ITransactionsDetailsRequest } from "../../models/api/chrysalis/ITransactionsDetailsRequest";
 import { ITransactionsDetailsResponse } from "../../models/api/chrysalis/ITransactionsDetailsResponse";
 import { ICachedTransaction } from "../../models/api/ICachedTransaction";
-import { ITransactionsDetailsRequest } from "../../models/api/ITransactionsDetailsRequest";
 import { ITransactionsCursor } from "../../models/api/og/ITransactionsCursor";
 import { TransactionsGetMode } from "../../models/api/og/transactionsGetMode";
 import { CHRYSALIS, OG } from "../../models/config/protocolVersion";
